+++ llm_guard/input_scanners/__init__.py
@@ -1,5 +1,6 @@
 """Input scanners init"""
 from .anonymize import Anonymize
+from .ban_code import BanCode
 from .ban_competitors import BanCompetitors
 from .ban_substrings import BanSubstrings
 from .ban_topics import BanTopics
@@ -17,6 +18,7 @@ from .util import get_scanner_by_name
 
 __all__ = [
     "Anonymize",
+    "BanCode",
     "BanCompetitors",
     "BanSubstrings",
     "BanTopics",