+++ llm_guard/input_scanners/util.py
@@ -1,6 +1,7 @@
 from typing import Dict, Optional
 
 from .anonymize import Anonymize
+from .ban_code import BanCode
 from .ban_competitors import BanCompetitors
 from .ban_substrings import BanSubstrings
 from .ban_topics import BanTopics
@@ -33,6 +34,9 @@ def get_scanner_by_name(scanner_name: str, scanner_config: Optional[Dict] = None
 
     if scanner_name == "Anonymize":
         return Anonymize(**scanner_config)
+
+    if scanner_name == "BanCode":
+        return BanCode(**scanner_config)
 
     if scanner_name == "BanCompetitors":
         return BanCompetitors(**scanner_config)