+++ docs/changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for `gunicorn` in the API deployment.
 - `NoRefusalLight` scanner that uses a common set of phrases to detect refusal as per research papers.
 - `Anonymize` and `Sensitive` scanners have a support of [lakshyakh93/deberta_finetuned_pii](https://huggingface.co/lakshyakh93/deberta_finetuned_pii) model.
+- `BanCode` scanner to detect and block code snippets in the prompt.
 
 ### Fixed
 - `InvisibleText` scanner to allow control characters like `\n`, `\t`, etc.